+++ src/main/java/com/example/studentservice/service/StudentService.java
@@ -1,5 +1,6 @@
 package com.example.studentservice.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,14 +9,14 @@ import com.example.studentservice.model.Student;
 public interface StudentService {
 	List<Student> findByFirstNameLike(String namePattern);
 	Student findByPassPortNumber(String passPortNumber);
-	List<Student> findByGradePointsBetween(Double start, Double end);
-	List<Student> findByGradePointsGreaterThanEqual(Double gradePoint);
-	List<Student> findTop5ByGradePoints(Double gradePoint);
+	List<Student> findByGradePointsBetween(BigDecimal start, BigDecimal end);
+	List<Student> findByGradePointsGreaterThanEqual(BigDecimal gradePoint);
+	List<Student> findTop5ByGradePoints(BigDecimal gradePoint);
 	List<Student> findByPassportNumberStartingWith(String prefix);
-	Student findFirstByGradePointsOrderByGradePointsDesc(Double gradePoint);
+	Student findFirstByGradePointsOrderByGradePointsDesc(BigDecimal gradePoint);
 	Student getStudentByStudentId(Integer studentId);
 	List<Student> getStudentByFirstName(String studentName);
-	List<Student> getStudentByGradePoints(Double gradePoints);
+	List<Student> getStudentByGradePoints(BigDecimal gradePoints);
 	List<Student> getStudentFirstNameLike(String namePattern);
 	void deleteStudent(Integer studentId);
 	Student getStudentByFirstNameAndLastName(String firstName, String lastName);
